sql: remove MEM_Static flag
This patch removes the MEM_Static flag. This flag was used to indicate
that a variable length MEM value will not change during the lifetime of
the MEM, so there is no need to manage the memory where the value is
located. Since this flag was introduced as an counterpart for ephemeral
and dynamic types, it is not really needed.

Additionally, this patch reworks the way you set static, ephemeral, and
dynamic MEM memory management types. Previously there were functions for
each MEM value type and each MEM memory management type, although some
functions were never used. Now value setting and memory management are
separated. For example, to set the value of a dynamic string in MEM, we
must set the value with mem_set_str() and have MEM manage memory with
mem_set_dynamic().

Part of #6051

NO_DOC=refactoring
NO_TEST=refactoring
NO_CHANGELOG=refactoring