sql: rework flag MEM_Ephem of struct MEM
This patch reworks the use of MEM_Ephem and replaces the flag with the
new is_ephemeral field.

Prior to this patch MEM_Ephem was used in two cases:
1) when a MEM was to contain a variable length value whose memory is
managed by another MEM;
2) when a MEM was to contain a variable length value whose memory is
managed elsewhere, which usually means that the memory allocated before
the MEM was created and freed after the MEM was destroyed.

In the second case, the MEM_Static flag was already used quite often
instead of MEM_Ephem. The mem_copy() function is the only one whose
behavior differs depending on whether the MEM is static or ephemeral,
but since it is never used to materialize an ephemeral MEM, there is no
difference between static and ephemeral MEM in the second case. This
means that in the second case we can mark MEM as static.

Part of #6051

NO_DOC=refactoring
NO_TEST=refactoring
NO_CHANGELOG=refactoring