sql: rework MEM_Number, MEM_Scalar and MEM_Any
This patch removes field flags from struct Mem, since this field is only
used to indicate that the type of the MEM is NUMBER, SCALAR, ANY, or the
same as the type of the value it contains. Instead, a new enum field has
been added.

Part of #6051

NO_DOC=refactoring
NO_TEST=refactoring
NO_CHANGELOG=refactoring