fio: fix default mode for files created by open()
For an unknown reason default fio.open() mode used to be set as
0777 & umask that basically equals 18, meaning no read/write/execute
rights for anyone.

Now the default is 0666 before umask, meaning read and write access.

Closes #7981

NO_DOC=bugfix

(cherry picked from commit 3d6025dcf9e7ef6e55123e5239086de46c28fce6)

----

Motivation for backporting to `release/2.11` (a future `2.11.10`
release) is that `fio.open` without the `mode` argument can be used in
tests and, in fact, I did that in `box-luatest/xlog_test.lua`. The test
was backported to `release/2.11` (PR #12979) and CI doesn't fail, while
a local run fails. I guess that CI runs tests from `root`.

Aside of that, the bugfix is quite safe to backport and it seems worth
to backport instead of patching the test (and taking the 2.11 vs 3.x
difference into account when writing other tests that are to be
backported).