cmake: always define TARGET_OS_* to 0 or 1
There is a system include file TargetConditionals.h on macOS, which
defines TARGET_OS_LINUX (and others) to 0 or 1. On the other side,
TARGET_OS_LINUX is also defined by trivia/config.h.cmake, but there it
has another possible values: undefined or 1. This inconsistency causes
issues like #8445, when TARGET_OS_LINUX is defined (to 0) in one file
and undefined in another. Let's always define it to 0 or 1.

Closes #8445

NO_DOC=bugfix