box: apply dynamic cfg even if option value is unchanged
If a configuration option value passed to `box.cfg` is the same as the
old one, the option handler isn't called. As a result, one can't restart
the server IPROTO connection or replication by passing the same URI to
`box.cfg`. This is inconvenient, for example:
 - To update an SSL key file without renaming it, one has to first pass
   an empty URI to `box.cfg.listen` and then reset it back.
 - To restart broken replication after fixing it manually (e.g. deleting
   a conflicting record on the replica), one has to first set
   `box.cfg.replication` to `{}` and then set it back instead of just
   calling `box.cfg{replication = box.cfg.replication}`.

There's no reason to skip an option update if the value is unchanged.
It should be up to the option configuration callback.

The only callback that may actually fail on an attempt to set the same
value is `box.cfg.memtx_memory`. This happens because the value is
rounded up to a multiple of the quota unit size so we just need to fix
the check accordingly.

Note that the `box.cfg.replication` configuration callback won't restart
replication if it's up and running even if the URI list is rearranged
thanks to commit 5994892c2f2c ("replication: fix replica disconnect upon
reconfiguration"). This commit doesn't break this behavior - it just
removes the `load_cfg.lua` code that would skip option update if the old
and new option values are equal in Lua.

Needed for tarantool/tarantool-ee#432
Closes #8551

NO_DOC=bug fix; this behavior isn't described anywhere in the docs