sql: separate PSEUDO cursor from sorter
Prior to this patch, the sorter used a PSEUDO cursor to pass data to
OP_Column where the data was properly prepared in the first OP_Column
for each row. This doesn't really make sense since we can prepare the
data before we pass it to OP_Column and we don't need the PRESUDO cursor
for that. This patch forces the sorter to properly prepare data before
passing it directly to OP_Column.

Needed for #8099

NO_DOC=internal
NO_TEST=internal
NO_CHANGELOG=internal