sql: unset nullRow for PSEUDO cursor
Prior to this patch, the nullRow field was set on the PSEUDO cursor
because at one place the opcode OP_NullRow was used to reset the cursor
state. This patch introduces the opcode OP_PseudoData which is used to
reset the state, so now nullRow is only used for cursors with no data,
as it should be. Also, the OP_PseudoData opcode is used to properly
prepare the data before passing it to the OP_Column, instead of
preparing the data in the first OP_Column for each row.

Needed for #8099

NO_DOC=internal
NO_TEST=internal
NO_CHANGELOG=internal