applier: do not pass event to ballot update trigger
Ballot watcher used to pass a boolean as an event to ballot_update
triggers, with "false" meaning there was some error while waiting for
ballot and one has to check diag for the error.

This was intended as a guard against stale diag contents (diag might
contain some leftovers from previous errors).

Actually, ballot watcher fiber clears the diag every time it handles an
error by itself, so the diag isn't empty if and only if the error wasn't
handled. One may check for `diag_is_empty()` instead of `success`
everywhere in triggers, so the parameter is extraneous. Remove it.

In-scope-of #8509

NO_DOC=refactoring
NO_TEST=refactoring
NO_CHANGELOG=refactoring