ci: fix PR update timestamp setter
The setup-python action defaults to searching for
a dependency file (requirements.txt or pyproject.toml
for pip, Pipfile.lock for pipenv or poetry.lock for
poetry) in the repository, and uses its hash as a
part of the cache key. We don't have one, so the
caching option is disabled.

NO_DOC=Workflow fix
NO_TEST=Workflow fix
NO_CHANGELOG=Workflow fix