relay: refactor is_raft_enabled flag
Relay had a is_raft_enabled member with mixed meaning: firstly, it was set
to true only when relay was ready to accept messages via cbus, and
secondly, it was set to true only for replicas which need raft updates
(newer than Tarantool 2.6.0 and not anonymous).

Let's better use the flag only as an indication that the relay is ready
to accept cbus pushes, and check whether the relay needs raft updates
separately.

The flag will be reused in the following commit, which will make tx
check that relay is connected prior to sending a message to it.

Prerequisite #7991

NO_DOC=refactoring
NO_TEST=refactoring
NO_CHANGELOG=refactoring