test: regenerate snapshots for #7974
The snapshots for #7974 tests use an old schema, 2.10 (the schema was
upgraded after they were generated), and in scope of #7149 DDL operations
with an old schema version were disabled: the problem is luatest needs to
grant permissions (which is a DDL operation) to run tests (it does it in
the test server instance source file here). As a result, the test fails
while trying to start a test server instance.

To fix this issue, we need to regenerate the snapshots files with
`box.schema.user.grant('guest', 'super')`.

While we are here, fix the instructions for generating snapshots for #7974
to refer to 2.10 source code.

Closes #8702

NO_CHANGELOG=<test fix>
NO_DOC=<test fix>