box: allow to set *_uuid options to NULL
This patch allows to set replicaset_uuid and instance_uuid to box.NULL.
This fixes the issue described in #8714, however it introduces another
change in behavior - we can now set these parameters to NULL even if
they weren't NULL before. However, since we still cannot set a different
uuid after setting the parameters to NULL, and we can still set the old
uuid for them, this behavior is considered acceptable.

Closes #8714

NO_DOC=bugfix
NO_CHANGELOG=the bug was not released