config: add logger wrapper with JSON encoding
The future config module works with hierarchical data a lot and needs a
way to show a table as a part a log message. The config module issues
log messages at different log levels and it is undesirable to perform
JSON encoding for messages that will be filtered out by the current log
level.

It is complicated to achieve the goal using the `log` module. Let's add
a wrapper as the temporary solution. The wrapper can be eliminated after
solving the problems in the `log` module that are linked in the
wrapper's code.

The wrapper also simplifies debugging of the config module itself.

NO_DOC=the module is for internal use from the config code
NO_CHANGELOG=see NO_DOC