ci: use workaround for LuaJIT profiling tests in all workflows
After this patch, the workaround for LuaJIT profiling tests to
avoid runner's shutdown due to no space left on the disk is going
to be used in all relevant workflows. Previously, we applied it
only for the coverage test, but we noticed that other jobs may be
affected as well after tarantool/tarantool#8737 is merged.

Follows up tarantool/tarantool#7472

NO_DOC=ci
NO_TEST=ci
NO_CHANGELOG=ci