net.box: export method constants to Lua automatically
Duplicating the net.box method constant enum in C and Lua is
inconvenient and error-prone. Let's export it to Lua automatically,
like we do with IPROTO constants. This will simplify addition of
new net.box methods (for example, WATCH_ONCE).

While we are at it,
 - move NETBOX_SELECT_WITH_POS closer to NETBOX_SELECT because they are
   related so better be defined together;
 - replace spaces with tabs in the encode/decode method tables for
   consistency.

Needed for #6493

NO_DOC=refactoring
NO_CHANGELOG=refactoring