box: save MsgPack encoding of original (user-provided) format clause
In scope of #4693 we need to save the MsgPack encoding of the original
(i.e, user-provided) format clause in the tuple format structure for
serialization to Lua and IPROTO (#8147, #8633): since tuple formats for
spaces are created from space definitions, we need to also save the MsgPack
encoding in the latter.

Since we need to pass the format clause MsgPack encoding to the runtime
tuple format constructor, refactor it to perform field definition decoding
in-place. We cannot use default field definition array decoding for
external formats (67578d1), so introduce a new field name decoder for this
case.

Needed for #4693

NO_CHANGELOG=refactoring
NO_DOC=refactoring
NO_TEST=refactoring