config/schema: add <schema object>:set()
It allows to construct a hierarchical data step-by-step or modify it
easily. Just like a field assignment operation for a flat data.

The function validates the value that is to be assigned, so data
constructed using a sequence of `:set()` calls is valid.

It performs the same path validation as `<schema object>:get()` to prevent
indexing of a scalar value or assigning an unknown record's field.

Part of #8725

NO_DOC=the module is for internal use from the config code
NO_CHANGELOG=see NO_DOC