core/lua: introduce func_adapter for Lua
We are going to introduce universal triggers registry and use it for
tarantool triggers (e.g. space.on_replace). However, when trigger is set
to the registry, we do not know which arguments will be passed there
from C. That is why current implementation of triggers in tarantool is not
suitable for use in universal registry.

We decided to introduce func_adapter interface which allows to call
functions from any programming language. Now we support only one
language, which is Lua, so for the most part this interface is needed
to invert dependencies.

The patch introduces func_adapter interface and its implementation for
Lua functions.

Part of #8657

NO_CHANGELOG=internal
NO_DOC=internal