config/schema: add <schema object>:filter()
It allows to walk over the hierarchical data and process it with
knowledge about corresponding schema nodes.

In conjunction with user-provided annotations (stored in the schema
nodes) it offers a powerful way to slice the data in different ways.

Part of #8725

NO_DOC=the module is for internal use from the config code
NO_CHANGELOG=see NO_DOC