config/schema: add <schema object>:apply_default()
This method adds default values from the `default` annotation for missed
fields. The annotation is supported for scalar values. The default can
be conditional, it is controlled by the `apply_default_if` annotation.

Part of #8725

NO_DOC=the module is for internal use from the config code
NO_CHANGELOG=see NO_DOC