config/schema: schema.set()
This is a shortcut for an array of unique string values from the given
list of allowed values. Useful for schema nodes declarations that
describes permissions such as {'read'}, {'read', 'write', 'execute'}
and so on.

Part of #8725

NO_DOC=the module is for internal use from the config code
NO_CHANGELOG=see NO_DOC