config/schema: add string-and-number scalar type
It is useful for options like `log_level` (with values like `5` or
`'info'`) or `synchro_quorum` (with values like `5` or `'N / 2 + 1'`).

This is a temporary solution. I want to introduce an explicit union
schema node later.

Fixes #8725

NO_DOC=the module is for internal use from the config code
NO_CHANGELOG=see NO_DOC