box: make fselect return a multi-line string instead of a table
The fselect space/index method returns select results in a pretty format
that looks like a table in the console, which encodes the results in the
YAML format. Before the yaml_pretty_multiline compat option was added,
multi-line strings emitted by the console looked mangled so the fselect
method returns a table of strings instead of a multi-line string for the
output to look good. Moreover, it also prepends the strings with the
special zero-width space character (U+200B) to employ a bug in the YAML
formatter: this character isn't printed but it makes the YAML formatter
emit the string without quotes. This hack won't work when #8756 is fixed
because we'll update the YAML formatter to output all non-printable
character as code points.

Now, that we have yaml_pretty_multiline compat option available and
enabled by default, we don't need these hacks anymore. Let's patch
fselect to return the result as a multi-line string and drop the
'use_nbsp' option, which was used to enable or disable the zero-width
space character hack. Let's also drop the 'print' option because one can
now print the fselect results directly.

Needed for #8756

NO_DOC=fselect is undocumented