config: introduce box_cfg applier
This patch introduces the box_cfg applier. This applier is used to
process box options. This appliers is called after mkdir applier because
we need all necessary directories to exist by the time the box options
are processed.

Part of #8789

NO_DOC=will be added later
NO_CHANGELOG=will be added later